LayoutTests/fast/forms/date/input-date-validation-message-expected.txt

Test for validationMessage IDL attribute for <input type=date>

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


No message
PASS testIt("", "", "") is ""
Value missing
FAIL testIt("", "", "") should be value missing. Was Fill out this field.
Type mismatch
PASS testIt("foo", "", "") is ""
Range overflow
FAIL testIt("1982-11-02", "", "1970-12-31") should be range overflow. Was Value must be less than or equal to 1970-12-31.
Range underflow
FAIL testIt("1982-11-02", "1990-05-25", "1990-12-24") should be range underflow. Was Value must be greater than or equal to 1990-05-25.
Step mismatch
FAIL testIt("1982-11-02", "", "", "123") should be step mismatch. Was Enter a valid value.
PASS successfullyParsed is true

TEST COMPLETE
